add : playground/list 페이지 수정건

+++ src/hooks/useCustomhook.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { useLocation, useNavigate, useParams } from "react-router-dom";
-import { deleteFont, updateFont } from "../api/FontsService";
+import { deleteFont, updateFont } from "../api/fontsService";
 import toast from "react-hot-toast";
 import { login, logout, signup } from "../api/FirebaseAPI";
 
@@ -12,8 +12,6 @@ export default function useCustomhook(){
     const [ bool, setBool ] = useState<boolean>(false);
     // 문자열타입
     const [ tempA, setTempA ] = useState<string>("");
-    const [ tempB, setTempB ] = useState<string>("");
-    const [ tempC, setTempC ] = useState<string>("");
     // 모든타입
     const [ item, setItem ] = useState<any>(null);
 
@@ -35,6 +33,8 @@ export default function useCustomhook(){
     }
     // 폰트 삭제 핸들러
     const handleDelete = async() => {
+      const isConfirmed = window.confirm("커스텀 폰트를 삭제하시겠습니까?");
+      if(!isConfirmed) return;
       try{
         await deleteFont(id!);
         toast.success(`삭제에 성공하였습니다.`);
@@ -64,25 +64,46 @@ export default function useCustomhook(){
     // 유저 로그인
     const handleLogin = async() => {
       const { email, password } = userData;
+      
+      // 폼검증
+      if (!email || !password) {
+        toast.error("이메일과 비밀번호를 입력해주세요.");
+        return;
+      }
+
+      if (!email.includes("@")) {
+        toast.error("이메일 형식이 올바르지 않습니다.");
+        return;
+      }
       try{
         await login(email, password); // tempA: email tempB: pw로 지정..
         toast.success("로그인 성공.");
         navigate("/");
       }catch(error:any){
-        if (error.code === "auth/user-not-found") toast.error("해당 이메일의 계정이 없습니다.");
-        else if (error.code === "auth/invalid-email") toast.error("해당 이메일의 계정이 없습니다.")
-        else if (error.code === "auth/wrong-password") toast.error("비밀번호가 잘못되었습니다.");
+        toast.error("이메일 또는 비밀번호가 올바르지 않습니다.");
       }
     }
     // 유저 회원가입
     const handleSignup = async () => {
       const { email, password, passwordCheck, username } = userData;
       
+    // --폼검증
+    // 모두 빈칸?
     if (!email || !password || !username) {
       toast.error("모든 항목을 입력해주세요.");
       return;
     }
-
+    // 이메일 형식 불일치
+    if(!email.includes("@")) {
+      toast.error("이메일 형식이 올바르지 않습니다..");
+      return;
+    }
+    // 비밀번호 길이제한
+    if(password.length < 6){
+      toast.error("비밀번호는 6자이상이여야 합니다.");
+      return;
+    }
+    // 패스워드 일치여부
     if (password !== passwordCheck) {
       toast.error("비밀번호가 일치하지 않습니다.");
       return;
@@ -93,9 +114,11 @@ export default function useCustomhook(){
       toast.success("회원가입 성공!");
       navigate("/user/login");
     } catch (err: any) {
-      if (err.code === "auth/invalid-email") toast.error("이메일 형식이 올바르지 않습니다.");
-      else if (err.code === "auth/email-already-in-use") toast.error("이미 사용 중인 이메일입니다.");
-      else toast.error("회원가입 실패");
+      if (err.code === "auth/email-already-in-use") {
+      toast.error("이미 사용 중인 이메일입니다.");
+      } else {
+        toast.error("회원가입 실패");
+      }
     }
     };
 
@@ -128,8 +151,6 @@ export default function useCustomhook(){
       bool,
       setBool,
       setTempA,
-      setTempB,
-      setTempC,
       userData,
 
       // 이벤트 핸들러
